feat: Add Flask web interface for OI Tracker
This commit introduces a complete web-based user interface for the OI Tracker application, built with Flask.

The new web interface provides a more user-friendly and secure way to interact with the application. Users can now enter their API credentials through a web form, and the application handles the authentication flow with Upstox.

The core OI tracking logic has been refactored into a separate module (`tracker_logic.py`) to improve code organization.

The web application includes:
- A home page for entering API credentials (`templates/index.html`).
- A dashboard page to display the OI data (`templates/dashboard.html`).
- A Flask server to handle web requests and authentication (`app.py`).
