+++ templates/index.html
@@ -1,0 +1,33 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Upstox OI Tracker - Login</title>
+    <style>
+        body { font-family: sans-serif; background-color: #f4f4f9; color: #333; display: flex; justify-content: center; align-items: center; height: 100vh; margin: 0; }
+        .container { background: white; padding: 2rem; border-radius: 8px; box-shadow: 0 4px 6px rgba(0,0,0,0.1); }
+        h1 { color: #0052cc; }
+        form { display: flex; flex-direction: column; }
+        label { margin-bottom: 0.5rem; font-weight: bold; }
+        input[type="text"], input[type="password"] { padding: 0.5rem; margin-bottom: 1rem; border: 1px solid #ccc; border-radius: 4px; }
+        input[type="submit"] { background-color: #0052cc; color: white; padding: 0.7rem; border: none; border-radius: 4px; cursor: pointer; font-size: 1rem; }
+        input[type="submit"]:hover { background-color: #003d99; }
+    </style>
+</head>
+<body>
+    <div class="container">
+        <h1>Upstox OI Tracker</h1>
+        <p>Please enter your API credentials to begin.</p>
+        <form action="/login" method="post">
+            <label for="api_key">API Key:</label>
+            <input type="text" id="api_key" name="api_key" required>
+
+            <label for="api_secret">API Secret:</label>
+            <input type="password" id="api_secret" name="api_secret" required>
+
+            <input type="submit" value="Login with Upstox">
+        </form>
+    </div>
+</body>
+</html>
